# apps/web-v2/src/lib/generated/collections.openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "CurateDS.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/collections": {
      "get": {
        "tags": [
          "CurateDS.Api"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CollectionResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CurateDS.Api"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/collections/{collectionId}/saved-views": {
      "get": {
        "tags": [
          "CurateDS.Api"
        ],
        "parameters": [
          {
            "name": "collectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SavedViewResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CurateDS.Api"
        ],
        "parameters": [
          {
            "name": "collectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSavedViewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedViewResponse"
                }
              }
            }
          }
        }
      }
    },
    "/collections/{collectionId}/saved-views/{viewId}": {
      "delete": {
        "tags": [
          "CurateDS.Api"
        ],
        "parameters": [
          {
            "name": "collectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "viewId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CollectionResponse": {
        "required": [
          "id",
          "name",
          "createdUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "createdUtc": {
            "type": "string",
            "format": "date-time"
          },
          "category": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "coverImageUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "color": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "CreateCollectionRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "category": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "coverImageUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "color": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SavedViewResponse": {
        "required": [
          "id",
          "collectionId",
          "name",
          "filtersJson",
          "createdUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "collectionId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "filtersJson": {
            "type": "string"
          },
          "createdUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateSavedViewRequest": {
        "required": [
          "name",
          "filtersJson"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "filtersJson": {
            "type": "string"
          }
        }
      },
...
    }
  }
}